Move Wear ApplicationModule to the commons lib
